improve seperation of concern in export function
